Moved auth function to util
